refactor : change package name entity to domain
