+++ src/main/java/racing/domain/Car.java
@@ -1,4 +1,4 @@
-package racing.entity;
+package racing.domain;
 
 public class Car {
     private static final int MAX_NAME_LENGTH = 5;
