+++ src/main/java/racing/domain/RacingCars.java
@@ -1,4 +1,4 @@
-package racing.entity;
+package racing.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
