+++ src/main/java/racing/generator/CarListGenerator.java
@@ -1,6 +1,6 @@
 package racing.generator;
 
-import racing.entity.Car;
+import racing.domain.Car;
 
 import java.util.ArrayList;
 import java.util.List;
